Fix CService constructor ambiguity in GetEffectiveTorProxy()
Cast GetArg() return (int64_t) to unsigned short for the port
parameter to resolve overload ambiguity across all platforms.

+++ src/tor/onion_v3.cpp
@@ -93,7 +93,7 @@ static std::string GetEffectiveTorProxy()
     }
 
     if (mapArgs.count("-tor") && mapArgs["-tor"] != "0") {
-        CService torProxy(mapArgs["-tor"], GetArg("-torsocks", 19099));
+        CService torProxy(mapArgs["-tor"], (unsigned short)GetArg("-torsocks", 19099));
         if (torProxy.IsValid()) {
             return torProxy.ToStringIPPort();
         }
